Separate reused helper functions close #67
